Override SublimeLinter.finalize_cmd method to reset command
SublimeLinter automatically appends the filename to the end of the
command, but unfortunately, golangci-lint doesn’t supports linting
individual files, or at least not very well. When the linter is executed
using a single file, packages are not scanned, which translates into
useless warnings about missing functions and global variables.