Add temporary directory handling to improve linter accuracy
When the user modifies a file but doesn’t saves the changes, the linter
is unable to report warnings and errors because golangci-lint is still
scanning the unmodified version of the code. From now on, the linter
will copy the project into a temporary directory, then take the editor
buffer as the content for the file that’s being modified. Then execute
the command against the temporary folder.