Fix output parser when golangci-lint errors in /dev/stderr
When a GOROOT file like `GOROOT/src/time/time.go` is analyzed, the
linter errors because some linters like “nolint” are unable to create
the AST cache. The errors are sent to /dev/stderr in non-JSON format
which makes it impossible for the SublimeLinter plugin to parse. Because
of this, I have opted to silence these warnings.