Simplify the 'GolangCILint' class
When using a recent version of 'golangci-lint', the generated output is
predictable enough to be parsed by a regex instead of converting JSON output.
This simplifies the code needed for linting considerably and removes the need of
a temporary file. Furthermore, by using the 'filename' capture group introduced
in SublimeLinter 4.13, a lot of false positives are fixed.